Initial checkin of Thomas Sailers code
I've tried to find the SCM for this soundmodem code, but since the
project seems to be dead for many years, the page has been removed, and
no new versions is up and running. The only way to find the source was
from a version tarball

This is threfore a direct import of the 0.18 tarball (with github
versions of LICENSE and README.md, instead of COPYING and README).

The code was released as GPLv2, which is a great license to keep it in.
That makes it possible to continue the great work of Thomas Sailer, and
hopefully ressurect the project at github.

This is still one of the best performing ax25-modems available, both
software and hardware.

Signed-off-by: Max Sikstrom <msikstrom@op5.com>